+++ google-cloud-bigquerystorage/src/main/java/com/google/cloud/bigquery/storage/v1/ConnectionWorkerPool.java
@@ -58,6 +58,11 @@ public class ConnectionWorkerPool {
   private final long maxInflightBytes;
 
   /*
+   * Max retry duration for retryable errors.
+   */
+  private final java.time.Duration maxRetryDuration;
+
+  /*
    * Behavior when inflight queue is exceeded. Only supports Block or Throw, default is Block.
    */
   private final FlowController.LimitExceededBehavior limitExceededBehavior;
@@ -191,11 +196,13 @@ public class ConnectionWorkerPool {
   public ConnectionWorkerPool(
       long maxInflightRequests,
       long maxInflightBytes,
+      java.time.Duration maxRetryDuration,
       FlowController.LimitExceededBehavior limitExceededBehavior,
       String traceId,
       BigQueryWriteSettings clientSettings) {
     this.maxInflightRequests = maxInflightRequests;
     this.maxInflightBytes = maxInflightBytes;
+    this.maxRetryDuration = maxRetryDuration;
     this.limitExceededBehavior = limitExceededBehavior;
     this.traceId = traceId;
     this.clientSettings = clientSettings;
@@ -348,6 +355,7 @@ public class ConnectionWorkerPool {
             writeSchema,
             maxInflightRequests,
             maxInflightBytes,
+            maxRetryDuration,
             limitExceededBehavior,
             traceId,
             clientSettings);