+++ google-cloud-bigquerystorage/src/test/java/com/google/cloud/bigquery/storage/v1/ConnectionWorkerPoolTest.java
@@ -152,7 +152,8 @@ public class ConnectionWorkerPoolTest {
             .setMaxConnectionsPerRegion(maxConnections)
             .build());
     ConnectionWorkerPool connectionWorkerPool =
-        createConnectionWorkerPool(maxRequests, /*maxBytes=*/ 100000);
+        createConnectionWorkerPool(
+            maxRequests, /*maxBytes=*/ 100000, java.time.Duration.ofSeconds(5));
 
     // Sets the sleep time to simulate requests stuck in connection.
     testBigQueryWrite.setResponseSleep(Duration.ofMillis(50L));
@@ -205,7 +206,8 @@ public class ConnectionWorkerPoolTest {
     ConnectionWorkerPool.setOptions(
         Settings.builder().setMaxConnectionsPerRegion(10).setMinConnectionsPerRegion(5).build());
     ConnectionWorkerPool connectionWorkerPool =
-        createConnectionWorkerPool(/*maxRequests=*/ 3, /*maxBytes=*/ 1000);
+        createConnectionWorkerPool(
+            /*maxRequests=*/ 3, /*maxBytes=*/ 1000, java.time.Duration.ofSeconds(5));
 
     // Sets the sleep time to simulate requests stuck in connection.
     testBigQueryWrite.setResponseSleep(Duration.ofMillis(50L));
@@ -254,7 +256,8 @@ public class ConnectionWorkerPoolTest {
     ConnectionWorkerPool.setOptions(
         Settings.builder().setMaxConnectionsPerRegion(10).setMinConnectionsPerRegion(5).build());
     ConnectionWorkerPool connectionWorkerPool =
-        createConnectionWorkerPool(/*maxRequests=*/ 3, /*maxBytes=*/ 100000);
+        createConnectionWorkerPool(
+            /*maxRequests=*/ 3, /*maxBytes=*/ 100000, java.time.Duration.ofSeconds(5));
 
     // Sets the sleep time to simulate requests stuck in connection.
     testBigQueryWrite.setResponseSleep(Duration.ofMillis(50L));
@@ -419,11 +422,13 @@ public class ConnectionWorkerPoolTest {
     return rowsBuilder.build();
   }
 
-  ConnectionWorkerPool createConnectionWorkerPool(long maxRequests, long maxBytes) {
+  ConnectionWorkerPool createConnectionWorkerPool(
+      long maxRequests, long maxBytes, java.time.Duration maxRetryDuration) {
     ConnectionWorkerPool.enableTestingLogic();
     return new ConnectionWorkerPool(
         maxRequests,
         maxBytes,
+        maxRetryDuration,
         FlowController.LimitExceededBehavior.Block,
         TEST_TRACE_ID,
         clientSettings);